Correct logic error in colloscope constraints building
A constraint was added "group in slot needs students" to avoid a problem where
groups could be marked in spot without the corresponding dynamic students.

However, the constraint was wrong. It enforced at least one dynamic student for each group and did not account
for the possibility of having students assigned to different slots.

The problem does not appear when there is at least one fixed student in a group because the "one interrogation per period" constraint
will avoid this issue.

So, remove the bad constraint and put in its place a new constraint "at most one interrogation per period" for empty groups. This should
solve the problem entirely.